+++ src/Cloud.Framework.Persistence.Abstractions/Sql.Base/DbQuery.cs
@@ -25,15 +25,21 @@ namespace Cloud.Framework.Persistence.Abstractions.Sql.Base
         public CommandType Type { get; }
 
         /// <summary>
+        /// The transaction that this query should be part of.
+        /// </summary>
+        public IDbTransaction? Transaction { get; }
+
+        /// <summary>
         /// The cancellation token for cancellable calls.
         /// </summary>
         public CancellationToken Token { get; }
 
-        private DbQuery(string sql, object? parameters = null, CancellationToken token = default, CommandType type = CommandType.Text) {
+        private DbQuery(string sql, object? parameters = null, CancellationToken token = default, CommandType type = CommandType.Text, IDbTransaction? transaction = null) {
             Sql = sql;
             Parameters = parameters;
             Token = token;
             Type = type;
+            Transaction = transaction;
         }
 
         /// <summary>
@@ -43,11 +49,12 @@ namespace Cloud.Framework.Persistence.Abstractions.Sql.Base
         /// <param name="parameters">The parameters used in the <paramref name="sql"/> statement.</param>
         /// <param name="token">The cancellation token for cancellable calls.</param>
         /// <param name="type">The <see cref="CommandType"/> for the <paramref name="sql"/> statement.</param>
+        /// <param name="transaction">The transaction to use if available.</param>
         /// <returns>A DbQuery object.</returns>
         /// <exception cref="ArgumentException">Thrown when the <paramref name="sql"/> argument is null or whitespace.</exception>
-        public static DbQuery Create(string sql, object? parameters = null, CancellationToken token = default, CommandType type = CommandType.Text) {
+        public static DbQuery Create(string sql, object? parameters = null, CancellationToken token = default, CommandType type = CommandType.Text, IDbTransaction? transaction = null) {
             if (string.IsNullOrWhiteSpace(sql)) throw new ArgumentException($"The {nameof(sql)} parameter can not be null or whitespace.", nameof(sql));
-            return new DbQuery(sql, parameters, token, type);
+            return new DbQuery(sql, parameters, token, type, transaction);
         }
     }
 }